feat(ena-deposition): refactor tests - split up liveness tests and helpers
