+++ ena-submission/src/ena_deposition/create_sample.py
@@ -374,6 +374,17 @@ def sample_table_handle_errors(
     return last_retry_time
 
 
+def create_sample_iter(
+    db_engine: Engine, config: Config, slack_config: SlackConfig, last_retry_time: datetime | None
+) -> datetime | None:
+    sync_state_with_submission_table(db_engine)
+
+    sample_table_create(db_engine, config)
+    # update submission_table state after creation
+    sync_state_with_submission_table(db_engine)
+    return sample_table_handle_errors(db_engine, config, slack_config, last_retry_time)
+
+
 def create_sample(config: Config, stop_event: threading.Event):
     db_engine = db_init(config.db_password, config.db_username, config.db_url)
     slack_config = slack_conn_init(
@@ -388,13 +399,10 @@ def create_sample(config: Config, stop_event: threading.Event):
             logger.warning("create_sample stopped due to exception in another task")
             return
         logger.debug("Checking for samples to create")
-        sync_state_with_submission_table(db_engine)
-
-        sample_table_create(db_engine, config)
-        sync_state_with_submission_table(db_engine)  # update submission_table state after creation
-        last_retry_time = sample_table_handle_errors(
-            db_engine, config, slack_config, last_retry_time
+        last_retry_time = create_sample_iter(
+            db_engine, config, slack_config, last_retry_time=last_retry_time
         )
+
         if stop_event.wait(timeout=config.time_between_iterations):
             logger.info("create_sample stopped due to exception in another task")
             return
